put solr_doc_builder and xml_solr_doc_builder in profiler module for namespacing
