+++ spec/unit/solr_doc_builder_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe SolrDocBuilder do
+RSpec.describe Profiler::SolrDocBuilder do
 
   before(:all) do
     @fake_druid = 'oo000oo0000'
@@ -17,7 +17,7 @@ RSpec.describe SolrDocBuilder do
     before(:each) do
       @hdor_client = double
       allow(@hdor_client).to receive(:mods).with(@fake_druid).and_return(@ng_mods_xml)
-      @doc_hash = SolrDocBuilder.new(@fake_druid, @hdor_client, nil).doc_hash
+      @doc_hash = Profiler::SolrDocBuilder.new(@fake_druid, @hdor_client, nil).doc_hash
     end
     it "id field set to druid" do
       expect(@doc_hash[:id]).to eq @fake_druid
@@ -37,8 +37,8 @@ RSpec.describe SolrDocBuilder do
       expect(@doc_hash).to include(:name_namePart_sim => ['Shindy'])
     end
     it "calls XmlSolrDocBuilder to populate hash fields from MODS" do
-      expect_any_instance_of(XmlSolrDocBuilder).to receive(:doc_hash).and_return([])
-      SolrDocBuilder.new(@fake_druid, @hdor_client, nil).doc_hash
+      expect_any_instance_of(Profiler::XmlSolrDocBuilder).to receive(:doc_hash).and_return([])
+      Profiler::SolrDocBuilder.new(@fake_druid, @hdor_client, nil).doc_hash
     end
   end
 
@@ -51,15 +51,15 @@ RSpec.describe SolrDocBuilder do
     context "smods_rec method (called in initialize method)" do
       it "returns Stanford::Mods::Record object" do
         expect(@real_hdor_client).to receive(:mods).with(@fake_druid).and_return(@ng_mods_xml)
-        sdb = SolrDocBuilder.new(@fake_druid, @real_hdor_client, nil)
+        sdb = Profiler::SolrDocBuilder.new(@fake_druid, @real_hdor_client, nil)
         expect(sdb.smods_rec).to be_an_instance_of(Stanford::Mods::Record)
       end
       it "raises exception if MODS xml for the druid is empty" do
         expect(@real_hdor_client).to receive(:mods).with(@fake_druid).and_return(Nokogiri::XML("<mods #{@ns_decl}/>"))
-        expect { SolrDocBuilder.new(@fake_druid, @real_hdor_client, nil) }.to raise_error(RuntimeError, Regexp.new("^Empty MODS metadata for #{@fake_druid}: <"))
+        expect { Profiler::SolrDocBuilder.new(@fake_druid, @real_hdor_client, nil) }.to raise_error(RuntimeError, Regexp.new("^Empty MODS metadata for #{@fake_druid}: <"))
       end
       it "raises exception if there is no MODS xml for the druid" do
-        expect { SolrDocBuilder.new(@fake_druid, @real_hdor_client, nil) }.to raise_error(Harvestdor::Errors::MissingMods)
+        expect { Profiler::SolrDocBuilder.new(@fake_druid, @real_hdor_client, nil) }.to raise_error(Harvestdor::Errors::MissingMods)
       end
     end
   end # context using Harvestdor::Client
